Fix minor issues in the ViewModel constructor injection code
