+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -55,9 +55,9 @@ import org.wordpress.android.ui.people.RoleChangeDialogFragment;
 import org.wordpress.android.ui.people.RoleSelectDialogFragment;
 import org.wordpress.android.ui.photopicker.PhotoPickerActivity;
 import org.wordpress.android.ui.plans.PlansActivity;
-import org.wordpress.android.viewmodel.PluginBrowserViewModel;
 import org.wordpress.android.ui.plugins.PluginBrowserActivity;
 import org.wordpress.android.ui.plugins.PluginDetailActivity;
+import org.wordpress.android.ui.plugins.PluginListFragment;
 import org.wordpress.android.ui.posts.AddCategoryFragment;
 import org.wordpress.android.ui.posts.EditPostActivity;
 import org.wordpress.android.ui.posts.EditPostSettingsFragment;
@@ -128,6 +128,7 @@ import dagger.android.support.AndroidSupportInjectionModule;
         LegacyModule.class,
         ReleaseToolsModule.class,
         AndroidSupportInjectionModule.class,
+        ViewModelModule.class,
         // Login flow library
         LoginAnalyticsModule.class,
         LoginFragmentModule.class,
@@ -256,8 +257,8 @@ public interface AppComponent extends AndroidInjector<WordPress> {
     void inject(HtmlToSpannedConverter object);
 
     void inject(PluginBrowserActivity object);
+    void inject(PluginListFragment object);
     void inject(PluginDetailActivity object);
-    void inject(PluginBrowserViewModel object);
 
     void inject(WordPressGlideModule object);
