+++ WordPress/src/main/java/org/wordpress/android/modules/ViewModelSubComponent.java
@@ -1,14 +1,13 @@
 package org.wordpress.android.modules;
 
 import org.wordpress.android.viewmodel.PluginBrowserViewModel;
-import org.wordpress.android.ui.stats.models.PostViewsModel;
 
 import dagger.Subcomponent;
 
 @Subcomponent
 public interface ViewModelSubComponent {
 
-    PluginBrowserViewModel pluginBrowserVM();
+    PluginBrowserViewModel pluginBrowserViewModel();
 
     @Subcomponent.Builder
     interface Builder {
