+++ WordPress/src/main/java/org/wordpress/android/ui/plugins/PluginListFragment.java
@@ -2,6 +2,7 @@ package org.wordpress.android.ui.plugins;
 
 
 import android.arch.lifecycle.Observer;
+import android.arch.lifecycle.ViewModelProvider;
 import android.arch.lifecycle.ViewModelProviders;
 import android.content.Context;
 import android.os.Bundle;
@@ -36,8 +37,14 @@ import org.wordpress.android.widgets.WPNetworkImageView;
 
 import java.util.List;
 
+import javax.inject.Inject;
+
 public class PluginListFragment extends Fragment {
     public static final String TAG = PluginListFragment.class.getName();
+
+    @Inject
+    ViewModelProvider.Factory mViewModelFactory;
+
     private static final String ARG_LIST_TYPE = "list_type";
 
     private PluginBrowserViewModel mViewModel;
@@ -57,8 +64,9 @@ public class PluginListFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        ((WordPress) getActivity().getApplication()).component().inject(this);
 
-        mViewModel = ViewModelProviders.of(getActivity()).get(PluginBrowserViewModel.class);
+        mViewModel = ViewModelProviders.of(getActivity(), mViewModelFactory).get(PluginBrowserViewModel.class);
         mListType = (PluginListType) getArguments().getSerializable(ARG_LIST_TYPE);
     }
